Rename index.js to app.js
