+++ src/server.js
@@ -1,6 +1,6 @@
 // Error tracking
 require('opbeat').start()
 
-const app = require('./index')
+const app = require('./app')
 
 app.listen(process.env.PORT, () => console.log('App listening on port ' + process.env.PORT))
